workspaces/workspace/lisa-leadflow-pro/SCRAPER/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.19.0
schedule>=1.1.0
pandas>=1.5.0
notion-client>=2.0.0
lxml>=4.9.0
